+++ crates/consensus/common/Cargo.toml
@@ -11,9 +11,6 @@ repository.workspace = true
 workspace = true
 
 [dependencies]
-# alloy
-alloy-primitives.workspace = true
-
 # reth
 reth-chainspec.workspace = true
 reth-consensus.workspace = true
