+++ crates/optimism/consensus/src/lib.rs
@@ -13,14 +13,14 @@ extern crate alloc;
 
 use alloc::{format, sync::Arc};
 use alloy_consensus::{BlockHeader as _, EMPTY_OMMER_ROOT_HASH};
-use alloy_primitives::{B256, B64};
+use alloy_primitives::B64;
 use core::fmt::Debug;
 use reth_chainspec::{EthChainSpec, EthereumHardforks};
 use reth_consensus::{Consensus, ConsensusError, FullConsensus, HeaderValidator};
 use reth_consensus_common::validation::{
     validate_against_parent_4844, validate_against_parent_eip1559_base_fee,
     validate_against_parent_hash_number, validate_against_parent_timestamp, validate_cancun_gas,
-    validate_header_base_fee, validate_header_extra_data, validate_header_gas, validate_state_root,
+    validate_header_base_fee, validate_header_extra_data, validate_header_gas,
 };
 use reth_execution_types::BlockExecutionResult;
 use reth_optimism_forks::OpHardforks;
@@ -204,9 +204,5 @@ impl<ChainSpec: EthChainSpec + OpHardforks, H: BlockHeader> HeaderValidator<H>
         }
 
         Ok(())
-    }
-
-    fn validate_state_root(&self, header: &H, root: B256) -> Result<(), ConsensusError> {
-        validate_state_root(header, root)
     }
 }
